apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "dependency-injector>=4.48.1",
    "instructor>=1.10.0",
    "openai>=1.98.0",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
]
